Use classpath normalization for the plugin directory input in tests
The plugin directory needs to be normalized as a classpath so that class timestamps and class order within the jar files are not considered when evaluating if the task is up to date. A regular binary comparison typically breaks the relocation of the cached artifacts.

Also, I added a comment explaining a previous change, in order to warn future maintaines about reverting it accidentally.