Build number fallback when product-info.json is missing
When configuring prepareSandbox we add some attributes to the jar
manifest, one of which being "Build-SDK" - the value of the attribute
is constructed from the contents of product-info.json (A metadata file located in the IDE directory).
However, when the target IDE is Android Studio, this file does not exist, and the value thereby becomes null;
this causes the following error:

Could not determine the dependencies of task ':prepareSandbox'.
> Failed to query the value of task ':prepareSandbox' property 'pluginJar'.
   > The value of a manifest attribute must not be null (Key=Build-SDK).

This commit makes it so that we fall back on the build number (also read
from the IDE directory) if product-info.json is missing.