Fix compiler resolution for long build numbers
In recent versions of Android Studio, the build number contains
additional components from major.minor.patch, e.g. 2020-3.1-patch-4 has
build number AI-203.7717.56.2031.7935034.
When configuration the plugin with a localPath that points to a recent Android Studio dist, these additional components causes the instrumentCode task to fail due to trying to resolve a non-existent compiler version (203.7717.56.2031.7935034).
With this commit we're stripping out components exceeding patch that are
not equal to SNAPSHOT or *. I.e, 1.2.3.4.5.SNAPSHOT becomes
1.2.3.SNAPSHOT.