Add `targetProductPart` property to allow installing plugin on the frontend part when running in Split Mode (#1563)
`targetProductPart` property is added to `IntelliJPlatformExtension` and `SplitModeAware`. It controls in which part of the IDE the developed plugin is installed. To do that, `PrepareSandboxTask` generates a properties file which configures the frontend process started from Split Mode. The path to the file is passed via `JETBRAINS_CLIENT_PROPERTIES` environment variable.