Ensure artifact transform failures are propagated
In case an artifact transform doesn't fail (aka throw a failure), Gradle will assume that it passed and never re-run it. The artifact transforms in this plugin swallow all exceptions, essentially causing builds to get into a failing state and making them hard to recover from that state.

This change makes sure we throw the exception when an artifact transform fails, so Gradle will re-run the artifact transform next time.